requirements.txt
future
emoji
beautifulsoup4
requests
sqlalchemy
python-telegram-bot
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
googletrans
jikanpy
speedtest-cli
coffeehouse
regex
wikipedia
telethon
spamwatch
alphabet_detector
nekos.py
fontTools
psutil
